Add files via upload
This is a project base on hospital-management-system which we can This project is based on a hospital management system that allows users to search for hospitals and perform CRUD (Create, Read, Update, Delete) operations. It features both UI and UX design elements to enhance user experience. Users can utilize GET, POST, PUT, and DELETE methods to interact with the system's data.each the hospital and get ,post,push,delete methods and as well .